Extract catalystType from CodecToEncoder
This extracts the Codec to catalystType logic to a separate object. This
is an preparation from for Spark 4 and was extracted from #16.

As part of this `nullable` can no longer be implemented using jvmType
and is therefore change to have more explicit logic that should have the
same results.
